ta3adod

+++ bonus/philosopher.h
@@ -6,7 +6,7 @@
 /*   By: houazzan <houazzan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/16 12:08:33 by houazzan          #+#    #+#             */
-/*   Updated: 2022/05/16 18:59:48 by houazzan         ###   ########.fr       */
+/*   Updated: 2022/05/17 15:44:49 by houazzan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ typedef struct s_philosopher
 typedef struct s_info
 {
 	int				philo_number;
-	int				time_to_eat;
+	long long		time_to_eat;
 	int				time_to_sleep;
 	int				time_to_die;
 	int				number_of_meals;
@@ -69,6 +69,6 @@ long long	time_diff(long long then, long long now);
 void		p_death( void *philo);
 void		exiting(t_info *rules, t_philosopher *philosopher);
 void		printing(t_info *rules, int id, char *str);
-void		sleep_time(int time, t_info *rules);
+void		sleep_time(long long time, t_info *rules);
 
 #endif
